Automatically add main url to host-map.php
If subsites is used in conjunction with staticpublishqueue, host-map.php currently has to be manually edited to include the main website's base URL every time a Subsite record is saved. Using these modules together already requires FilesystemPublisher::$static_base_url to be set, so we may as well make use of this setting if it is available to save this manual editing.