NEW revert button that matches active subsite to tab
The subsite ID is updated in the PHP session, and the JavaScript had no way of
reverting this back once if a user saw the modal on in their admin session. This
could lead to confusion about what to do next. Now there is a revert button that
will set the PHP session SubsiteID back to that of the tab beneath the modal,
making it safe to edit once again.

In order to support this, other tabs also need to update, and modals either displayed
or hidden, which was already a feature. However these changes only happen on remote tabs
(or Documents as per the API definition), so a new custom event has been introduced in
order to be able to have the current tab/document update itself - but without coupling
the React component to jQuery.entwine that handles the component lifecycle (externally
from React).