Trigger SubsiteID update on load, safely
Previous attempts to update the stored Subsite ID in the localStorage when a page loaded for the first time would cause an infinite refresh loop.
Reasons to set the ID on load are for hard refreshed tabs were a way to remove the modal blocking an edit, and could cause sync issues when multiple tabs are open - undermining the efforts recent commits have set to resolve.

This also solves an issue where loading a edit form for a page that did not exist on any other subsite would set the SubsiteID in the PHP Session backend, based on the subsite the page in question. This would create multiple tab sync issues in the CMS silently (no way to tell the UI to block an edit). By setting the SubsiteID when the page loads fresh we can avoid these situations.