Add tests for new SubsiteChangeAlert javascript
This lead to refactoring the SubsiteChangeAlert component to not rely
on global state of the DOM, and instead take props in and reference them
directly to perform its required tasks. Equally it accepts a callback for
delegating the revert task. This is cleaner and more testable.

Also fix small issue with adding divs for the modal as many times as it
appears.